# docs/api/openapi.yaml
openapi: 3.0.0
paths:
  /health:
    get:
      operationId: AppController_health
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - health
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a1
        - auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a1
  /auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags: *a1
  /users/me:
    get:
      operationId: UsersController_me
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /listings:
    get:
      operationId: ListingsController_list
      parameters: []
      responses:
        "200":
          description: ""
      tags: &a2
        - listings
    post:
      operationId: ListingsController_create
      parameters: []
      responses:
        "201":
          description: ""
      security:
        - bearer: []
      tags: *a2
  /listings/{id}:
    get:
      operationId: ListingsController_get
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a2
  /bookings:
    get:
      operationId: BookingsController_list
      parameters: []
      responses:
        "200":
          description: ""
      security: &a3
        - bearer: []
      tags: &a4
        - bookings
    post:
      operationId: BookingsController_create
      parameters: []
      responses:
        "201":
          description: ""
      security: *a3
      tags: *a4
  /bookings/{id}:
    get:
      operationId: BookingsController_get
      parameters: []
      responses:
        "200":
          description: ""
      security: *a3
      tags: *a4
  /bookings/{id}/confirm:
    post:
      operationId: BookingsController_confirm
      parameters: []
      responses:
        "201":
          description: ""
      security: *a3
      tags: *a4
info:
  title: Creators Marketplace API
  description: ""
  version: 0.1.0
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas: {}